+++ cli/abeamer-cli.ts
@@ -268,10 +268,14 @@ The commands are:
 
   function commandCreate(): void {
 
-    const projName = cmdParam;
+    const projName = fsix.toPosixSlash(cmdParam);
     if (projName === '' || projName[0] === '-'
       || projName.search(/[\\\/\?\*\+]/) !== -1) {
-      throw `Project name ${projName} is not valid`;
+      throw `Error: ${projName} is not valid project name`;
+    }
+
+    if (!sysFs.existsSync(projName)) {
+      throw `Error: Project ${projName} already exists`;
     }
 
     const ROOT_PATH = fsix.toPosixSlash(__dirname) + '/..';