+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Relaxes the `abeamer create` to support a project names with spaces and non-latin characters.
 - If `server-agent-puppeteer` has `page.goto` it logs the error.
 - (DEV)`gulp bump-version` checks for error conditions of badge creation.
+- `abeamer create` checks if the project name already exists, and throws error if is true.
 
 ## [0.2.6] -
 ### Fixed